examples/google_adk/french_translation_agent/pyproject.toml
[project]
name = "adk-french-translation-agent-example"
version = "0.1.0"
description = "French translation agent example using ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "google-genai>=1.9.0", # For Gemini model
    "google-adk>=1.0.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-sdk = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
